+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Initial project setup
 - Added config dependency
 - Added utils and debug functions
+- Added config handler